optimize the logic
